listview: use size_t, where applicable
Also store window sizes in unsigned integers, because it is assumed, that
getmaxy and getmaxx do not return negative values in these cases. This removes
sign compare warnings, when enabled.

+++ src/listview.h
@@ -4,15 +4,15 @@
 struct listview {
 	WINDOW *window;
 	struct listmodel *model;
-	unsigned int first;
-	unsigned int index;
+	size_t first;
+	size_t index;
 };
 
 void listview_up(struct listview *view);
 void listview_down(struct listview *view);
 void listview_pageup(struct listview *view);
 void listview_pagedown(struct listview *view);
-unsigned int listview_getindex(struct listview *view);
+size_t listview_getindex(struct listview *view);
 void listview_resize(struct listview *view, unsigned int width, unsigned int height);
 void listview_init(struct listview *view, struct listmodel *model, unsigned int x, unsigned int y, unsigned int width, unsigned int height);
 void listview_free(struct listview *view);
